Dockerfile to reproduce the issue (not a *good* Dockerfile!)
You may not want to actually merge this in b/c it re-clones the repository. That's not how you would normally write a Dockerfile. I wrote it that way to demonstrate that it performs all the same steps and gets the error "The name 'GeneratedClass' does not exist in the current context"